Add files via upload
Formatting in Excel means a trick that you can use to modify the data's appearance in a worksheet. You can format the data in various ways, like you can format the font of the cells or the table with the help of the styles and format tab present in the Home tab.